created a seperate extended sidebar component
